move tests and enable gtfn runner
