+++ tests/functional_tests/ffront_tests/test_execution.py
@@ -727,63 +727,3 @@ def test_tuple_with_local_field_in_reduction_shifted(reduction_setup):
     expected = red[rs.v2e_table][:, 0]
 
     assert np.allclose(expected, out)
-
-
-def test_tuple_program_return_constructed_inside(fieldview_backend):
-    size = 10
-    a = np_as_located_field(IDim)(np.ones((size,)))
-    b = np_as_located_field(IDim)(2 * np.ones((size,)))
-    out_a = np_as_located_field(IDim)(np.zeros((size,)))
-    out_b = np_as_located_field(IDim)(np.zeros((size,)))
-
-    @field_operator
-    def pack_tuple(
-        a: Field[[IDim], float64], b: Field[[IDim], float64]
-    ) -> tuple[Field[[IDim], float64], Field[[IDim], float64]]:
-        return (a, b)
-
-    @program(backend=fieldview_backend)
-    def prog(
-        a: Field[[IDim], float64],
-        b: Field[[IDim], float64],
-        out_a: Field[[IDim], float64],
-        out_b: Field[[IDim], float64],
-    ):
-        pack_tuple(a, b, out=(out_a, out_b))
-
-    prog(a, b, out_a, out_b, offset_provider={})
-
-    assert np.allclose(a, out_a)
-    assert np.allclose(b, out_b)
-
-
-def test_tuple_program_return_constructed_inside_nested(fieldview_backend):
-    size = 10
-    a = np_as_located_field(IDim)(np.ones((size,)))
-    b = np_as_located_field(IDim)(2 * np.ones((size,)))
-    c = np_as_located_field(IDim)(3 * np.ones((size,)))
-    out_a = np_as_located_field(IDim)(np.zeros((size,)))
-    out_b = np_as_located_field(IDim)(np.zeros((size,)))
-    out_c = np_as_located_field(IDim)(np.zeros((size,)))
-
-    @field_operator
-    def pack_tuple(
-        a: Field[[IDim], float64], b: Field[[IDim], float64], c: Field[[IDim], float64]
-    ) -> tuple[tuple[Field[[IDim], float64], Field[[IDim], float64]], Field[[IDim], float64]]:
-        return ((a, b), c)
-
-    @program(backend=fieldview_backend)
-    def prog(
-        a: Field[[IDim], float64],
-        b: Field[[IDim], float64],
-        c: Field[[IDim], float64],
-        out_a: Field[[IDim], float64],
-        out_b: Field[[IDim], float64],
-        out_c: Field[[IDim], float64],
-    ):
-        pack_tuple(a, b, c, out=((out_a, out_b), out_c))
-
-    prog(a, b, c, out_a, out_b, out_c, offset_provider={})
-
-    assert np.allclose(a, out_a)
-    assert np.allclose(b, out_b)
